add book via tportal done

+++ book/models.py
@@ -9,11 +9,12 @@ class BookTypes(models.Model):
 
 class Books(models.Model):
     bgrade=models.ForeignKey(BookTypes,on_delete=models.CASCADE)
+    #bgrade=models.IntegerField(default=10)
     name=models.CharField(max_length=200)
     subject=models.CharField(max_length=200,default="10science,11maths,12english,Main,Neet...")
     details=models.CharField(max_length=200)
     review=models.CharField(max_length=200)
-    rating=models.IntegerField(default=0)
+    rating=models.FloatField(default=0)
     image=models.ImageField(default="images/tst2.jpg")
     author=models.CharField(max_length=200,default="NCERT")
     file=models.FileField(default="images/default.pdf")
